Upgrade from angular 1.1.5 to 1.2.32
The 1.1 series was (intentionally?) unstable and had breaking changes with eatch patch version.
Because 1.1.x is unstable, 1.1.5 it is no longer available from bower, and so it's impossible to do
a repeatable build from it.

The 1.2 series is the stable series based on the growth of 1.1. 1.2.32 is the most recent,
highly-patched (and thus most-stable) version of 1.2, which has no breaking changes for patch updates.
So 1.2.32 was the safest choice to make the JS build repeatable.

Changes that affected this code from 1.1 to 1.2
(see https://docs.angularjs.org/guide/migration#migrating-from-1-0-to-1-2 for the full list):
- ngRoute moved to its own module
- ngBindHtmlUnsafe is now just ngBindHtml, and requires the ngSanitize module